Address Feedback
Merge Test files

group these files together:
-  test_unwind_errors_path_syntax.py + test_unwind_errors_preserve_and_fields.py
-   test_unwind_dotted_path.py + test_unwind_dotted_path_options.py
 -  test_unwind_include_array_index.py +  test_unwind_index_field_name_acceptance.py
- test_unwind_scalar_passthrough.py + test_unwind_unicode.py
-  test_unwind_mixed_shape.py + test_unwind_nested_arrays.py  + test_unwind_multi_stage.py
