+++ documentdb_tests/compatibility/tests/core/operator/stages/unwind/test_unwind_multi_doc_and_stages.py
@@ -1,4 +1,4 @@
-"""Tests for $unwind stage — mixed-shape multi-document behavior."""
+"""Tests for $unwind stage — multi-document, nested arrays, and multi-stage chaining."""
 
 from __future__ import annotations
 
@@ -154,11 +154,123 @@ UNWIND_MIXED_SHAPE_TESTS: list[StageTestCase] = [
     ),
 ]
 
+# Property [Nested Arrays]: $unwind peels exactly one level of array nesting
+# per invocation - inner arrays are preserved as elements, and successive
+# $unwind stages on the same field flatten additional nesting levels.
+UNWIND_NESTED_ARRAYS_TESTS: list[StageTestCase] = [
+    StageTestCase(
+        "nested_array_of_arrays",
+        docs=[{"_id": 1, "a": [[1, 2], [3]]}],
+        pipeline=[{"$unwind": {"path": "$a"}}],
+        expected=[
+            {"_id": 1, "a": [1, 2]},
+            {"_id": 1, "a": [3]},
+        ],
+        msg="$unwind should produce one document per inner array",
+    ),
+    StageTestCase(
+        "nested_deeply_nested",
+        docs=[{"_id": 1, "a": [[[1]], [[2, 3]]]}],
+        pipeline=[{"$unwind": {"path": "$a"}}],
+        expected=[
+            {"_id": 1, "a": [[1]]},
+            {"_id": 1, "a": [[2, 3]]},
+        ],
+        msg="$unwind should peel only one level, preserving deeper nesting",
+    ),
+    StageTestCase(
+        "nested_mixed_scalars_and_arrays",
+        docs=[{"_id": 1, "a": [1, [2, 3], 4]}],
+        pipeline=[{"$unwind": {"path": "$a"}}],
+        expected=[
+            {"_id": 1, "a": 1},
+            {"_id": 1, "a": [2, 3]},
+            {"_id": 1, "a": 4},
+        ],
+        msg="$unwind should preserve inner arrays alongside scalar elements",
+    ),
+    StageTestCase(
+        "nested_successive_unwind_flattens",
+        docs=[{"_id": 1, "a": [[10, 20], [30]]}],
+        pipeline=[{"$unwind": {"path": "$a"}}, {"$unwind": {"path": "$a"}}],
+        expected=[
+            {"_id": 1, "a": 10},
+            {"_id": 1, "a": 20},
+            {"_id": 1, "a": 30},
+        ],
+        msg="Successive $unwind stages on the same field should flatten additional nesting",
+    ),
+]
+
+# Property [Multi-Stage Unwind]: chaining multiple $unwind stages composes
+# correctly with independent state tracking, cross-product semantics, and
+# preserve interactions.
+UNWIND_MULTI_STAGE_TESTS: list[StageTestCase] = [
+    StageTestCase(
+        "multi_stage_index_resets_per_subarray",
+        docs=[{"_id": 1, "a": [[10, 20], [30]]}],
+        pipeline=[
+            {"$unwind": {"path": "$a", "includeArrayIndex": "i1"}},
+            {"$unwind": {"path": "$a", "includeArrayIndex": "i2"}},
+        ],
+        expected=[
+            {"_id": 1, "a": 10, "i1": INT64_ZERO, "i2": INT64_ZERO},
+            {"_id": 1, "a": 20, "i1": INT64_ZERO, "i2": Int64(1)},
+            {"_id": 1, "a": 30, "i1": Int64(1), "i2": INT64_ZERO},
+        ],
+        msg=(
+            "Second $unwind includeArrayIndex should reset to 0 for each"
+            " sub-array produced by the first $unwind"
+        ),
+    ),
+    StageTestCase(
+        "multi_stage_cross_product_different_fields",
+        docs=[{"_id": 1, "a": [1, 2], "b": ["x", "y"]}],
+        pipeline=[{"$unwind": {"path": "$a"}}, {"$unwind": {"path": "$b"}}],
+        expected=[
+            {"_id": 1, "a": 1, "b": "x"},
+            {"_id": 1, "a": 1, "b": "y"},
+            {"_id": 1, "a": 2, "b": "x"},
+            {"_id": 1, "a": 2, "b": "y"},
+        ],
+        msg=(
+            "Two $unwind stages on different array fields should produce"
+            " a cross product of elements"
+        ),
+    ),
+    StageTestCase(
+        "multi_stage_preserve_first_filter_second",
+        docs=[
+            {"_id": 1, "a": [1, 2], "b": ["x"]},
+            {"_id": 2, "a": [3], "b": None},
+            {"_id": 3, "a": [4]},
+            {"_id": 4, "a": None, "b": ["z"]},
+        ],
+        pipeline=[
+            {"$unwind": {"path": "$a", "preserveNullAndEmptyArrays": True}},
+            {"$unwind": {"path": "$b", "preserveNullAndEmptyArrays": False}},
+        ],
+        expected=[
+            {"_id": 1, "a": 1, "b": "x"},
+            {"_id": 1, "a": 2, "b": "x"},
+            {"_id": 4, "a": None, "b": "z"},
+        ],
+        msg=(
+            "preserveNullAndEmptyArrays on first $unwind followed by false"
+            " on second should filter documents where second path is null or missing"
+        ),
+    ),
+]
+
+UNWIND_COMPLEX_TESTS = (
+    UNWIND_MIXED_SHAPE_TESTS + UNWIND_NESTED_ARRAYS_TESTS + UNWIND_MULTI_STAGE_TESTS
+)
+
 
 @pytest.mark.aggregate
-@pytest.mark.parametrize("test_case", pytest_params(UNWIND_MIXED_SHAPE_TESTS))
-def test_unwind_mixed_shape(collection, test_case: StageTestCase):
-    """Test $unwind mixed-shape multi-document behavior."""
+@pytest.mark.parametrize("test_case", pytest_params(UNWIND_COMPLEX_TESTS))
+def test_unwind_complex(collection, test_case: StageTestCase):
+    """Test $unwind multi-document, nested arrays, and multi-stage chaining."""
     populate_collection(collection, test_case)
     result = execute_command(
         collection,
